# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "devstack",
    "description": "Local development orchestration with automatic port allocation, structured logs, and file watching",
    "license": {
      "name": "MIT"
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/v1/sources": {
      "get": {
        "tags": [
          "daemon"
        ],
        "operationId": "list_sources",
        "responses": {
          "200": {
            "description": "List registered log sources",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "daemon"
        ],
        "operationId": "add_source",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Source added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddSourceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources/{name}": {
      "get": {
        "tags": [
          "daemon"
        ],
        "operationId": "source_status",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Source name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Source indexing status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceStatusResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "daemon"
        ],
        "operationId": "remove_source",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Source name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Source removed"
          }
        }
      }
    },
    "/v1/sources/{name}/logs": {
      "get": {
        "tags": [
          "daemon"
        ],
        "operationId": "source_logs_view",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Source name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Combined source log view",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogViewResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddSourceRequest": {
        "type": "object",
        "required": [
          "name",
          "paths"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "retention": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AddSourceResponse": {
        "type": "object",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SourceSummary"
          }
        }
      },
...
      "FacetFilter": {
        "type": "object",
        "required": [
          "field",
          "kind",
          "values"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetValueCount"
            }
          }
        }
      },
      "FacetValueCount": {
        "type": "object",
        "required": [
          "value",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "LogEntry": {
        "type": "object",
        "required": [
          "ts",
          "service",
          "stream",
          "level",
          "message",
          "raw"
        ],
        "properties": {
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "json": {
            "nullable": true
          },
          "level": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "raw": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "stream": {
            "type": "string"
          },
          "ts": {
            "type": "string"
          }
        }
      },
...
      "LogViewResponse": {
        "type": "object",
        "required": [
          "entries",
          "truncated",
          "total",
          "filters"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogEntry"
            }
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetFilter"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "total_exact": {
            "type": "boolean"
          },
          "truncated": {
            "type": "boolean"
          }
        }
      },
...
      "SourceStatusResponse": {
        "type": "object",
        "required": [
          "source"
        ],
        "properties": {
          "index": {
            "allOf": [
              {
                "$ref": "#/components/schemas/crate.sources.SourceIndexState"
              }
            ],
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/SourceSummary"
          }
        }
      },
      "SourceSummary": {
        "type": "object",
        "required": [
          "name",
          "paths",
          "created_at",
          "effective_retention_seconds"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "effective_retention_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "retention_seconds": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "minimum": 0
          }
        }
      },
      "SourcesResponse": {
        "type": "object",
        "required": [
          "sources"
        ],
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceSummary"
            }
          }
        }
      },
...
    }
  },
...
}